Track IP addresses using ping (as a proof of concept)
You can now also specify IP address, not just MAC address. The plugin will automatically recognize it as IP and use a special tracker module to ping it every 30 seconds (for now hard coded timing). 
This is mainly to demonstrate the easiness of adding new methods for detection.
Also renamed a lot of modules and functions to better reflect their meaning (it is no longer just routers and macs ;-)