Work-around for problem with Debian 11 (Bullseye) end python 3.9
Under investigation (Help in finding the true issue will be appreciated!): 
After installing Debian Bullseye (which includes python 3.9) the plugin would generate lots of communication errors when using ssh tracker. Upgrading to the latest beta of Domoticz does not remedy the situation. The errors would only occur when starting Domoticz. Not when (quickly) restarting the plugin from within an already running Domoticz instance. So -after not being able to find the cause and a real solution- I build a little delay into the plugin startup. It will wait for the third heartbeat before actually starting the plugin code.